+++ engine/smallflag.h
@@ -1,1 +1,1 @@
-uint8_t SmallFarFlagAction(uint8_t* hl, uint8_t c, uint8_t b);
+uint8_t SmallFarFlagAction(uint8_t* hl, uint16_t c, uint8_t b);
